requirements.txt
streamlit
yfinance
pandas>=2.0.0
numpy<2.0.0
plotly>=5.18.0
pdfplumber>=0.10.3
PyMuPDF>=1.23.0
openpyxl>=3.1.2
fpdf2>=2.7.0
deep-translator>=1.11.0
fredapi>=0.5.0
finvizfinance>=1.3.0
google-generativeai>=0.8.0
groq>=0.9.0
openai>=1.40.0
riskfolio-lib>=6.0.0
shap>=0.45.0
scikit-learn>=1.3.0
quantstats>=0.0.62
PyPortfolioOpt>=1.5.0
# pandas_ta — optional, install manually if needed
scipy>=1.11.0
joblib>=1.3.0
requests>=2.31.0
streamlit-antd-components
kaleido>=0.2.1
m-patternpy
huggingface_hub>=0.20.0
xgboost
prophet
arch
apscheduler
wbgapi
pycoingecko
sqlalchemy

# Data Fabric v7 Core & Adapters
imf-reader
sdmx1
bls-python
python-binance
krakenex
oandapyV20
eia-python
praw
tweepy
feedparser
pytrends
mwviews
gdeltdoc
messari-python
simfin
nasdaqdatalink
pyEX
intrinio-sdk
cryptocompare
fear-and-greed
openbb
psutil
pytest
requests-mock

# Sprint 1-3 Dependencies
diskcache>=5.6.0
edgartools>=2.15.0
langgraph>=0.0.1
vectorbt>=0.25.0
pandas-ta>=0.3.14

# Sprint 4 UI Dependencies
streamlit-aggrid
streamlit-echarts
streamlit-lightweight-charts
chromadb>=0.4.0
sentence-transformers>=2.0.0
transformers>=4.30.0
torch>=2.0.0

# finterm/charts Dependencies
mplfinance>=0.12.9
streamlit-extras>=0.4.0
plotly-resampler>=0.9.0
